Палитра языков вынесена из карты форков
- Добавлен модуль src/utils/lang.ts с LANG_COLORS, LANG_SHORT и langColor
- Убран лишний импорт useForkMap в резюме и ветках пулл-реквестов: сетевой код карты больше не попадает в чужие чанки
- Обновлены все потребители палитры
