+++ src/composables/useForkMap.ts
@@ -46,22 +46,6 @@ export interface MapStats {
   languages: number
 }
 
-export const LANG_COLORS: Record<string, string> = {
-  'C#': '#4caf50',
-  Vue: '#42b883',
-  JavaScript: '#f1e05a',
-  HTML: '#e34c26',
-  Go: '#00add8',
-  Python: '#3572a5',
-  SCSS: '#c6538c',
-  PowerShell: '#5391fe',
-  прочее: '#9a9a9a',
-}
-
-export const LANG_SHORT: Record<string, string> = {
-  JavaScript: 'JS',
-}
-
 export const OWNER = 'MaxNagibator'
 
 export type SizeBy = 'stars' | 'forks' | 'merged' | 'size' | 'commits'
@@ -85,8 +69,6 @@ const LANG_ORDER = [
   'PowerShell',
   'прочее',
 ]
-
-export const langColor = (lang: string): string => LANG_COLORS[lang] ?? '#9a9a9a'
 
 export const DOMAINS: Domain[] = [
   { key: 'apps', label: 'приложения', width: 66, bisector: -90 },
